GC and performance improvements
- Limit direct use of org.objectweb.asm.Type to reduce GC pressure
- Inline reference access for property calls
- Add line info for export/import entries
- Change module semantics to use LinkedHashMap to retain insertion order
- Fix string concat evaluation bug and add invokedynamic
- Improve best fit matcher performance
- Improve internal error creation performance
- Add fast path to update own, writable properties

+++ src/main/java/com/github/anba/es6draft/runtime/objects/binary/TypedArrayObject.java
@@ -121,7 +121,7 @@ public final class TypedArrayObject extends IntegerIndexedObject implements Arra
         /* step 12 */
         int elementSize = elementType.size();
         /* step 13 */
-        long indexedPosition = (long) ((index * elementSize) + offset);
+        long indexedPosition = (index * elementSize) + offset;
         /* step 15 */
         return GetValueFromBuffer(buffer, indexedPosition, elementType);
     }
@@ -156,7 +156,7 @@ public final class TypedArrayObject extends IntegerIndexedObject implements Arra
         /* step 14 */
         int elementSize = elementType.size();
         /* step 15 */
-        long indexedPosition = (long) ((index * elementSize) + offset);
+        long indexedPosition = (index * elementSize) + offset;
         /* steps 17-18 */
         SetValueInBuffer(buffer, indexedPosition, elementType, numValue);
         /* step 19 */
